+++ Packages/com.unity.inputsystem/InputSystem/Editor/Internal/EditorHelpers.cs
@@ -65,23 +65,13 @@ namespace UnityEngine.InputSystem.Editor
             CheckOut(path);
         }
 
-        public static string ReadAllText(string path)
+        public static string GetPhysicalPath(string path)
         {
-            // Note that FileUtil.GetPhysicalPath(string) is only available in 2021.2 or newer
+            // Note that we can only get physical path for 2021.2 or newer
 #if UNITY_2021_2_OR_NEWER
-            return File.ReadAllText(FileUtil.GetPhysicalPath(path));
+            return FileUtil.GetPhysicalPath(path);
 #else
-            return File.ReadAllText(path);
-#endif
-        }
-
-        public static void WriteAllText(string path, string contents)
-        {
-            // Note that FileUtil.GetPhysicalPath(string) is only available in 2021.2 or newer
-#if UNITY_2021_2_OR_NEWER
-            File.WriteAllText(path: FileUtil.GetPhysicalPath(path), contents: contents);
-#else
-            File.WriteAllText(path: path, contents: contents);
+            return path;
 #endif
         }